+++ spec/analytics.spec.ts
@@ -32,7 +32,7 @@ describe('Analytics', function() {
                 return Promise.resolve({
                     pages: {
                         previous: null,
-                        next: '/v1/apps/1337/analytics/subjects/summary?after=PAGE_TWO'
+                        next: '/v2/apps/1337/analytics/subjects/summary?after=PAGE_TWO'
                     },
                     data: (new Array(10)).fill(fakeSubjectStats)
                 });
@@ -41,9 +41,9 @@ describe('Analytics', function() {
 
         function _validate(options) {
             if (options.url.indexOf('PAGE_TWO') > 0) {
-                expect(options.url).to.equal('/v1/apps/1337/analytics/subjects/summary?after=PAGE_TWO');
+                expect(options.url).to.equal('/v2/apps/1337/analytics/subjects/summary?after=PAGE_TWO');
             } else {
-                expect(options.url).to.equal('/v1/apps/1337/analytics/subjects/summary');
+                expect(options.url).to.equal('/v2/apps/1337/analytics/subjects/summary');
             }
             expect(options.headers).to.be.an('object');
         }
@@ -67,7 +67,7 @@ describe('Analytics', function() {
         }
 
         function _validate(options) {
-            expect(options.url).to.equal('/v1/apps/1337/analytics/metrics/keys');
+            expect(options.url).to.equal('/v2/apps/1337/analytics/metrics/keys');
         }
 
         const result = await bup.analytics.getAllMetricKeys({ _payload, _validate });