+++ spec/criteria.spec.ts
@@ -23,7 +23,7 @@ describe('criterion', function() {
         }
 
         function _validate(options) {
-            expect(options.url).to.equal(`/v1/apps/1337/criteria/${criterion.id}`);
+            expect(options.url).to.equal(`/v2/apps/1337/criteria/${criterion.id}`);
             expect(options.method).to.be.oneOf([undefined, 'GET']);
             expect(options.headers).to.be.an('object');
         }
@@ -50,7 +50,7 @@ describe('criterion', function() {
                 return Promise.resolve({
                     pages: {
                         previous: null,
-                        next: '/v1/apps/1337/criteria?after=PAGE_TWO'
+                        next: '/v2/apps/1337/criteria?after=PAGE_TWO'
                     },
                     data: (new Array(10)).fill(criterion)
                 });
@@ -59,9 +59,9 @@ describe('criterion', function() {
 
         function _validate(options) {
             if (options.url.indexOf('PAGE_TWO') > 0) {
-                expect(options.url).to.equal('/v1/apps/1337/criteria?after=PAGE_TWO');
+                expect(options.url).to.equal('/v2/apps/1337/criteria?after=PAGE_TWO');
             } else {
-                expect(options.url).to.equal('/v1/apps/1337/criteria');
+                expect(options.url).to.equal('/v2/apps/1337/criteria');
             }
             expect(options.headers).to.be.an('object');
         }
@@ -80,7 +80,7 @@ describe('criterion', function() {
         }
 
         function _validate(options) {
-            expect(options.url).to.equal(`/v1/apps/1337/criteria`);
+            expect(options.url).to.equal(`/v2/apps/1337/criteria`);
             expect(options.method).to.equal('POST');
             expect(options.headers).to.be.an('object');
         }
@@ -97,7 +97,7 @@ describe('criterion', function() {
         }
 
         function _validate(options) {
-            expect(options.url).to.equal(`/v1/apps/1337/criteria/${criterion.id}`);
+            expect(options.url).to.equal(`/v2/apps/1337/criteria/${criterion.id}`);
             expect(options.method).to.equal('DELETE');
             expect(options.headers).to.be.an('object');
         }