+++ spec/integration.spec.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { expect } from 'chai';
-import { Achievement, AchievementRequest, BadgeUp, Condition, CriterionType, EarnedAchievement, EventRequest, EventV1, EventV2Preview, Operation } from '../src';
+import { Achievement, AchievementRequest, BadgeUp, Condition, CriterionType, EarnedAchievement, EventRequest, Operation } from '../src';
 
 const INTEGRATION_API_KEY = process.env.INTEGRATION_API_KEY;
 
@@ -189,6 +189,9 @@ describe('integration tests', function() {
         const iterator = client.earnedAchievements.getIterator();
         for (const achievementPromise of iterator) {
             const achievement: EarnedAchievement | undefined = await achievementPromise;
+            if (!achievement) {
+                continue; // last page may be empty
+            }
             expect(achievement).to.be.an('object');
             expect(achievement!.achievementId).to.be.a('string');
         }
@@ -202,10 +205,13 @@ describe('integration tests', function() {
 
         const eventRequest = new EventRequest(subject, key, { '@inc': 5 });
 
-        const eventResponse: EventV1 = await client.events.create(eventRequest);
+        const eventResponse = await client.events.create(eventRequest);
         expect(eventResponse).to.be.an('object');
-        const event = eventResponse.event;
-        const progress = eventResponse.progress;
+        expect(eventResponse.results).to.be.an('array');
+        expect(eventResponse.results).to.have.length.greaterThan(0);
+
+        const event = eventResponse.results[0].event;
+        const progress = eventResponse.results[0].progress;
 
         expect(event).to.be.an('object');
         expect(event.key).to.be.equal(key);
@@ -261,28 +267,6 @@ describe('integration tests', function() {
         }
     });
 
-    it('should send an event and get progress back v2', async function() {
-        const client = new BadgeUp({ apiKey: INTEGRATION_API_KEY });
-
-        const subject = 'nodejs-ci-' + randomString();
-        const key = 'test';
-
-        const eventRequest = new EventRequest(subject, key, { '@inc': 5 });
-
-        const eventResponse: EventV2Preview = await client.events.createV2Preview(eventRequest);
-        expect(eventResponse).to.be.an('object');
-        expect(eventResponse.results).to.be.an('array');
-        expect(eventResponse.results).to.have.length.greaterThan(0);
-        eventResponse.results.forEach((e) => {
-            expect(e).to.be.an('object');
-            expect(e.event).to.be.an('object');
-            expect(e.event.applicationId).to.be.a('string');
-            expect(e.event.id).to.be.a('string');
-            expect(e.event.key).to.be.a('string');
-            expect(e.event.subject).to.be.a('string');
-        });
-    });
-
     it('should get achievement progress for a subject', async function() {
         const client = new BadgeUp({ apiKey: INTEGRATION_API_KEY });
 
@@ -291,7 +275,7 @@ describe('integration tests', function() {
 
         const eventRequest = new EventRequest(subject, key, { '@inc': 5 });
 
-        const eventResponse: EventV1 = await client.events.create(eventRequest);
+        const eventResponse = await client.events.create(eventRequest);
         expect(eventResponse).to.be.an('object'); // other tests check event response results
 
         const progressResponse = await client.progress.query().subject(subject).getAll();