+++ src/analytics/index.ts
@@ -29,7 +29,7 @@ export class AnalyticsResource {
         check.assert(check.integer(numDays) && check.greater(numDays, 0), 'numDays must be a positive integer');
 
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/events/last-n-days?duration=${numDays}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/events/last-n-days?duration=${numDays}`
         }, userOpts);
     }
 
@@ -43,7 +43,7 @@ export class AnalyticsResource {
         check.assert.string(subject, 'subject must be a string');
 
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/events/last-n-days/subject/${subject}?duration=${numDays}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/events/last-n-days/subject/${subject}?duration=${numDays}`
         }, userOpts);
     }
 
@@ -56,7 +56,7 @@ export class AnalyticsResource {
         check.assert(check.integer(numDays) && check.greater(numDays, 0), 'numDays must be a positive integer');
 
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/subjects/last-n-days?duration=${numDays}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/subjects/last-n-days?duration=${numDays}`
         }, userOpts);
     }
 
@@ -69,7 +69,7 @@ export class AnalyticsResource {
         check.assert(check.integer(numDays) && check.greater(numDays, 0), 'numDays must be a positive integer');
 
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/subjects/new/last-n-days?duration=${numDays}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/subjects/new/last-n-days?duration=${numDays}`
         }, userOpts);
     }
 
@@ -82,7 +82,7 @@ export class AnalyticsResource {
         check.assert(check.integer(numDays) && check.greater(numDays, 0), 'numDays must be a positive integer');
 
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/earnedachievements/last-n-days?duration=${numDays}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/earnedachievements/last-n-days?duration=${numDays}`
         }, userOpts);
     }
 
@@ -93,7 +93,7 @@ export class AnalyticsResource {
      */
     public *getSubjectsSummaryIterator(userOpts?) {
         const pageFn = () => {
-            let url = `/v1/apps/${this.context.applicationId}/${ENDPT}/subjects/summary`;
+            let url = `/v2/apps/${this.context.applicationId}/${ENDPT}/subjects/summary`;
             return () => {
                 return this.context.http.makeRequest({ url }, userOpts).then(function(body) {
                     url = body.pages.next;
@@ -112,7 +112,7 @@ export class AnalyticsResource {
      */
     public getAllMetricKeys(userOpts?) {
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/metrics/keys`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/metrics/keys`
         }, userOpts).then(obj => obj.data);
     }
 }