+++ src/apiKeys/index.ts
@@ -77,7 +77,7 @@ export class ApiKeysResource {
      */
     public listScopes(userOpts?): Promise<ApiKeyScope[]> {
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/scopes`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/scopes`
         }, userOpts).then((body) => body.data);
     }
 }