+++ src/events/index.ts
@@ -1,7 +1,6 @@
-import { defaultsDeep } from 'lodash';
 import { Common } from '../common';
 import { ResourceContext } from '../utils/ResourceContext';
-import { EventRequest, EventV1, EventV2Preview } from './Event.class';
+import { Event, EventRequest, EventResults } from './Event.class';
 
 const ENDPT = 'events';
 
@@ -9,7 +8,7 @@ const ENDPT = 'events';
  * Events resource
  */
 export class EventsResource {
-    private common: Common<EventV1>;
+    private common: Common<Event>;
 
     /**
      * Construct the Events resource
@@ -25,14 +24,7 @@ export class EventsResource {
      * @param userOpts option overrides for this request
      * @returns A promise that resolves to the provided event
      */
-    public create(object: EventRequest, userOpts?): Promise<EventV1> {
-        return this.common.create(object, userOpts);
+    public create(object: EventRequest, userOpts?): Promise<EventResults> {
+        return this.common.create<EventResults>(object, userOpts);
     }
-
-    public createV2Preview(object: EventRequest, userOpts?): Promise<EventV2Preview> {
-        // TODO: test how this works with user-provided headers, should be resolved by defaultsDeep
-        userOpts = defaultsDeep(userOpts, { headers: { 'X-V2-PREVIEW': 'true' } });
-        return this.common.create<EventV2Preview>(object, userOpts);
-    }
-
 }