+++ src/metrics/index.ts
@@ -59,7 +59,7 @@ export class MetricQueryBuilder {
 
         return this.context.http.makeRequest({
             method: 'DELETE',
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}?${querystring.stringify(queryBy)}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}?${querystring.stringify(queryBy)}`
         }, userOpts);
     }
 }
@@ -116,7 +116,7 @@ export class MetricsResource {
         check.assert.string(subject, 'subject must be a string');
 
         let array = [];
-        let url = `/v1/apps/${this.context.applicationId}/${ENDPT}/${subject}`;
+        let url = `/v2/apps/${this.context.applicationId}/${ENDPT}/${subject}`;
 
         const pageFn = () => {
             return this.context.http.makeRequest({ url }, userOpts).then(function(body) {
@@ -144,7 +144,7 @@ export class MetricsResource {
         check.assert.string(subject, 'subject must be a string');
 
         const pageFn = () => {
-            let url = `/v1/apps/${this.context.applicationId}/${ENDPT}/${subject}`;
+            let url = `/v2/apps/${this.context.applicationId}/${ENDPT}/${subject}`;
             return () => {
                 return this.context.http.makeRequest({ url }, userOpts).then(function(body) {
                     url = body.pages.next;
@@ -168,7 +168,7 @@ export class MetricsResource {
         check.assert.string(key, 'key must be a string');
 
         return this.context.http.makeRequest({
-            url: `/v1/apps/${this.context.applicationId}/${ENDPT}/${subject}/${key}`
+            url: `/v2/apps/${this.context.applicationId}/${ENDPT}/${subject}/${key}`
         }, userOpts);
     }